Move PlaylistTrsck->SSCE code to SpotifyPlaylistLoader
